move MotorTester outside
