+++ app/client/dashboard/components/__tests__/MotorTester-test.js
@@ -4,7 +4,7 @@ describe('MotorTester', function () {
   it('generates fake data', function () {
     var React = require('react/addons');
     var MotorTester = require('../MotorTester.js');
-    var RobotActions = require('../../../actions/RobotActions');
+    var RobotActions = require('../../actions/RobotActions');
     var TestUtils = React.addons.TestUtils;
 
     var tester = TestUtils.renderIntoDocument(
